+++ packages/saber-highlighter-prism/loadLanguages.js
@@ -1,103 +1,53 @@
-/* globals Prism */
-const components = require('prismjs/components.js')
+const Prism = require(`prismjs`)
+const prismComponents = require(`prismjs/components`)
 
-let peerDependentsMap = null
+// Get the real name of a language given it or an alias
+const getBaseLanguageName = (nameOrAlias, components = prismComponents) => {
+  if (components.languages[nameOrAlias]) {
+    return nameOrAlias
+  }
 
-function getPeerDependentsMap() {
-  const peerDependentsMap = {}
-  Object.keys(components.languages).forEach(language => {
-    if (language === 'meta') {
-      return false
+  return Object.keys(components.languages).find(language => {
+    const { alias } = components.languages[language]
+    if (!alias) return false
+    if (Array.isArray(alias)) {
+      return alias.includes(nameOrAlias)
     }
 
-    if (components.languages[language].peerDependencies) {
-      let { peerDependencies } = components.languages[language]
-      if (!Array.isArray(peerDependencies)) {
-        peerDependencies = [peerDependencies]
-      }
-
-      peerDependencies.forEach(peerDependency => {
-        if (!peerDependentsMap[peerDependency]) {
-          peerDependentsMap[peerDependency] = []
-        }
-
-        peerDependentsMap[peerDependency].push(language)
-      })
-    }
+    return alias === nameOrAlias
   })
-  return peerDependentsMap
 }
 
-function getPeerDependents(mainLanguage) {
-  if (!peerDependentsMap) {
-    peerDependentsMap = getPeerDependentsMap()
+module.exports = function loadPrismLanguage(language) {
+  const baseLanguage = getBaseLanguageName(language)
+
+  if (!baseLanguage) {
+    throw new Error(`Prism doesn't support language '${language}'.`)
   }
 
-  return peerDependentsMap[mainLanguage] || []
-}
-
-function loadLanguages(arr, withoutDependencies) {
-  // If no argument is passed, load all components
-  if (!arr) {
-    arr = Object.keys(components.languages).filter(language => {
-      return language !== 'meta'
-    })
-  }
-
-  if (arr && arr.length === 0) {
+  if (Prism.languages[baseLanguage]) {
+    // Don't load already loaded language
     return
   }
 
-  if (!Array.isArray(arr)) {
-    arr = [arr]
+  const languageData = prismComponents.languages[baseLanguage]
+
+  if (languageData.option === `default`) {
+    // Default language has already been loaded by Prism
+    return
   }
 
-  arr.forEach(language => {
-    if (!components.languages[language]) {
-      // Find the language by alias name
-      for (const name of Object.keys(components.languages)) {
-        const { alias } = components.languages[name]
-        const match = Array.isArray(alias)
-          ? alias.includes(language)
-          : alias === language
-        if (match) {
-          language = name
-        }
-      }
+  if (languageData.require) {
+    // Load the required language first
+    if (Array.isArray(languageData.require)) {
+      languageData.require.forEach(loadPrismLanguage)
+    } else {
+      loadPrismLanguage(languageData.require)
     }
+  }
 
-    if (!components.languages[language]) {
-      return
-    }
-
-    // Load dependencies first
-    if (!withoutDependencies && components.languages[language].require) {
-      loadLanguages(components.languages[language].require)
-    }
-
-    const pathToLanguage = 'prismjs/components/prism-' + language
-    delete require.cache[require.resolve(pathToLanguage)]
-    delete Prism.languages[language]
-    require(pathToLanguage)
-
-    // Reload dependents
-    const dependents = getPeerDependents(language).filter(dependent => {
-      // If dependent language was already loaded,
-      // we want to reload it.
-      if (Prism.languages[dependent]) {
-        delete Prism.languages[dependent]
-        return true
-      }
-
-      return false
-    })
-    if (dependents.length > 0) {
-      loadLanguages(dependents, true)
-    }
-  })
+  require(`prismjs/components/prism-${baseLanguage}.js`)
 }
 
-module.exports = function(arr) {
-  // Don't expose withoutDependencies
-  loadLanguages(arr)
-}
+/* Exposed for unit testing */
+module.exports.getBaseLanguageName = getBaseLanguageName