+++ packages/saber-plugin-pwa/lib/create-element.js
@@ -1,0 +1,8 @@
+const stringify = require('stringify-attributes')
+
+module.exports = (tagName, attrs, content) => {
+  const attrString = stringify(attrs)
+  return content
+    ? `<${tagName}${attrString}>${content}</${tagName}`
+    : `<${tagName}${attrString} />`
+}