+++ packages/saber-plugin-pwa/lib/get-app-config.js
@@ -1,0 +1,11 @@
+const defaults = {
+  name: 'Saber PWA App',
+  themeColor: '#ffffff',
+  assetsVersion: ''
+}
+
+module.exports = (config = {}) => {
+  return Object.assign({}, defaults, config, {
+    name: config.name || defaults.name
+  })
+}