+++ packages/saber/vue-renderer/lib/utils.js
@@ -1,0 +1,1 @@
+exports.prefixSpace = input => (input ? ` ${input}` : '')