+++ packages/saber-highlighter-prism/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/egoist/saber/compare/saber-highlighter-prism@0.2.2...saber-highlighter-prism@0.3.0) (2019-05-08)
+
+### Features
+
+- add saber-plugin-prismjs ([1bcf1e6](https://github.com/egoist/saber/commit/1bcf1e6))
+
 ## [0.2.2](https://github.com/egoist/saber/compare/saber-highlighter-prism@0.2.1...saber-highlighter-prism@0.2.2) (2019-05-03)
 
 ### Bug Fixes