+++ packages/saber-plugin-query-posts/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/egoist/saber/compare/saber-plugin-query-posts@0.2.0...saber-plugin-query-posts@0.3.0) (2019-05-11)
+
+### Bug Fixes
+
+- remove pages.extendPageProp ([5cafde8](https://github.com/egoist/saber/commit/5cafde8))
+
+### BREAKING CHANGES
+
+- `api.pages.extendPageProp` is removed, you should directly mutate `page` object instead
+
+# [0.2.0](https://github.com/egoist/saber/compare/saber-plugin-query-posts@0.1.5...saber-plugin-query-posts@0.2.0) (2019-05-08)
+
+### Features
+
+- i18n support ([#137](https://github.com/egoist/saber/issues/137)) ([1613626](https://github.com/egoist/saber/commit/1613626))
+
 ## [0.1.5](https://github.com/egoist/saber/compare/saber-plugin-query-posts@0.1.4...saber-plugin-query-posts@0.1.5) (2019-05-03)
 
 ### Bug Fixes